Finishing of inventory-system (in scene)
The now implemented inventory-system is not interactive but can be used in order to see what items the player has collected.

The player can not interact with the inventory yet. In order to do that, we would need another inventory system, which allows the player to interact with in only in specific circumstances (starting a puzzle)